backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
groq==0.4.2
langchain==0.1.9
langchain-groq==0.0.1
pymupdf==1.23.26
python-docx==1.1.0
faiss-cpu==1.7.4
pydantic==2.6.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.1
aiofiles==23.2.1 